builder: install 'rapstore-builder' service.
Install with

    fab builder.setup

After a reboot, the builder is not automatically restarted.
It can be started/stopped manually with

    fab builder.start

    fab builder.stop

Its executed as the 'builder' user with code install in ~builder.
Logs go to ~builder/logs.

Its run in a `tmux` session as not ready to go in docker.
And was lazy to use systemd for now.